refresh and add test
