+++ src/components/artist-question-screen/artist-question-screen.e2e.test.js
@@ -2,7 +2,30 @@ import React from "react";
 import {configure, shallow} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import ArtistQuestionScreen from "./artist-question-screen.jsx";
-import {questionArtist} from "../../mocks-test/questions";
+
+const mock = {
+  question: {
+    type: `artist`,
+    song: {
+      artist: ``,
+      src: ``
+    },
+    answers: [
+      {
+        artist: `one`,
+        picture: `pic-one`,
+      },
+      {
+        artist: `two`,
+        picture: `pic-two`,
+      },
+      {
+        artist: `three`,
+        picture: `pic-three`,
+      },
+    ],
+  }
+};
 
 configure({adapter: new Adapter()});
 
@@ -11,7 +34,7 @@ const mockEvent = {
 };
 
 it(`Click on user answer should pass to the callback data-object from which this answer was created`, () => {
-  const {question} = questionArtist;
+  const {question} = mock;
   const onAnswer = jest.fn();
   const userAnswer = {
     artist: `one`,
