+++ src/components/artist-question-screen/artist-question-srceen.test.js
@@ -1,10 +1,27 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import ArtistQuestionScreen from "./artist-question-screen.jsx";
-import {questionArtistType} from "../../types";
+
+const question = {
+  type: `artist`,
+  song: {
+    artist: `Jim Beam`,
+    src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
+  },
+  answers: [{
+    picture: `https://api.adorable.io/avatars/128/0`,
+    artist: `John Snow`,
+  }, {
+    picture: `https://api.adorable.io/avatars/128/1`,
+    artist: `Jack Daniels`,
+  }, {
+    picture: `https://api.adorable.io/avatars/128/2`,
+    artist: `Jim Beam`,
+  }],
+};
 
 it(`Should render ArtistQuestionScreen`, () => {
-  const tree = renderer.create(<ArtistQuestionScreen question={questionArtistType} onAnswer={() => {
+  const tree = renderer.create(<ArtistQuestionScreen question={question} onAnswer={() => {
   }}/>).toJSON();
   expect(tree).toMatchSnapshot();
 });
