+++ src/components/genre-question-screen/genre-question-screen.e2e.test.js
@@ -2,15 +2,40 @@ import React from "react";
 import {configure, shallow} from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import GenreQuestionScreen from "./genre-question-screen.jsx";
-import {questionGenre} from "../../mocks-test/questions";
 
 configure({adapter: new Adapter()});
 
+const mock = {
+  question: {
+    type: `genre`,
+    genre: `rock`,
+    answers: [
+      {
+        src: `path`,
+        genre: `rock`,
+      },
+      {
+        src: `path`,
+        genre: `jazz`,
+      },
+      {
+        src: `path`,
+        genre: `jazz`,
+      },
+      {
+        src: `path`,
+        genre: `blues`,
+      },
+    ],
+  },
+};
+
 it(`When user answers genre question form is not sent`, () => {
+  const {question} = mock;
   const onAnswer = jest.fn();
   const genreQuestion = shallow(<GenreQuestionScreen
     onAnswer={onAnswer}
-    question={questionGenre}
+    question={question}
   />);
 
   const form = genreQuestion.find(`form`);
@@ -21,27 +46,4 @@ it(`When user answers genre question form is not sent`, () => {
 
   expect(onAnswer).toHaveBeenCalledTimes(1);
   expect(formSendPrevention).toHaveBeenCalledTimes(1);
-});
-
-it(`User answer passed to callback is consistent with "userAnswer" prop`, () => {
-  const onAnswer = jest.fn((...args) => [...args]);
-  const userAnswer = [false, true, false, false];
-
-  const genreQuestion = shallow(<GenreQuestionScreen
-    onAnswer={onAnswer}
-    question={questionGenre}
-  />);
-
-  const form = genreQuestion.find(`form`);
-  const inputTwo = genreQuestion.find(`input`).at(1);
-
-  inputTwo.simulate(`change`, {target: {checked: true}});
-  form.simulate(`submit`, {preventDefault() {}});
-
-  expect(onAnswer).toHaveBeenCalledTimes(1);
-
-  expect(onAnswer.mock.calls[0][0]).toMatchObject(questionGenre);
-  expect(onAnswer.mock.calls[0][1]).toMatchObject(userAnswer);
-
-  expect(genreQuestion.find(`input`).map((it) => it.prop(`checked`))).toEqual(userAnswer);
 });
