+++ src/components/genre-question-screen/genre-question-screen.test.js
@@ -1,9 +1,26 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import GenreQuestionScreen from "./genre-question-screen.jsx";
-import {questionGenre} from "../../mocks-test/questions";
+
+const question = {
+  type: `genre`,
+  genre: `rock`,
+  answers: [{
+    src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
+    genre: `rock`,
+  }, {
+    src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
+    genre: `blues`,
+  }, {
+    src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
+    genre: `jazz`,
+  }, {
+    src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
+    genre: `rock`,
+  }],
+};
 
 it(`Should render GenreQuestionScreen`, () => {
-  const tree = renderer.create(<GenreQuestionScreen onAnswer={() => {}} question={questionGenre}/>).toJSON();
+  const tree = renderer.create(<GenreQuestionScreen onAnswer={() => {}} question={question}/>).toJSON();
   expect(tree).toMatchSnapshot();
 });
