--- src/mocks-test/questions.js
@@ -1,86 +1,0 @@
-const questionGenre = {
-  question: {
-    type: `genre`,
-    genre: `rock`,
-    answers: [
-      {
-        src: `path`,
-        genre: `rock`,
-      },
-      {
-        src: `path`,
-        genre: `jazz`,
-      },
-      {
-        src: `path`,
-        genre: `jazz`,
-      },
-      {
-        src: `path`,
-        genre: `blues`,
-      },
-    ],
-  },
-};
-
-const questionArtist = {
-  question: {
-    type: `artist`,
-    song: {
-      artist: ``,
-      src: ``
-    },
-    answers: [
-      {
-        artist: `one`,
-        picture: `pic-one`,
-      },
-      {
-        artist: `two`,
-        picture: `pic-two`,
-      },
-      {
-        artist: `three`,
-        picture: `pic-three`,
-      },
-    ],
-  }
-};
-
-const questions = [
-  {
-    type: `genre`,
-    genre: `rock`,
-    answers: [{
-      src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-      genre: `rock`,
-    }, {
-      src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-      genre: `blues`,
-    }, {
-      src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-      genre: `jazz`,
-    }, {
-      src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-      genre: `rock`,
-    }],
-  }, {
-    type: `artist`,
-    song: {
-      artist: `Jim Beam`,
-      src: `https://upload.wikimedia.org/wikipedia/commons/4/4e/BWV_543-fugue.ogg`,
-    },
-    answers: [{
-      picture: `https://api.adorable.io/avatars/128/1`,
-      artist: `John Snow`,
-    }, {
-      picture: `https://api.adorable.io/avatars/128/2`,
-      artist: `Jack Daniels`,
-    }, {
-      picture: `https://api.adorable.io/avatars/128/3`,
-      artist: `Jim Beam`,
-    }],
-  },
-];
-
-export {questionGenre, questionArtist, questions};
